feat: 태그 적용 및 오타 수정
