+++ src/components/commons/Card/TagList.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { useEffect, useLayoutEffect, useRef, useState } from 'react';
 import debounce from 'lodash.debounce';
