+++ src/components/products/mypage/Participating.tsx
@@ -1,4 +1,4 @@
-'use clients';
+'use client';
 
 import { useEffect, useState } from 'react';
 import Link from 'next/link';
