+++ src/hooks/useQueryTab.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { useRouter, useSearchParams } from 'next/navigation';
 
 export function useQueryTab<T extends string>(
