+++ src/types/tags.ts
@@ -52,3 +52,10 @@ export type BandSessionType =
   | 'STRING_INSTRUMENT'
   | 'PERCUSSION'
   | 'KEYBOARD';
+
+export type ParticipantStatus =
+  | 'PENDING'
+  | 'APPROVED'
+  | 'REJECTED'
+  | 'COMPLETED'
+  | 'CANCELED';
